\name{alphaMcmcPost-methods}
\alias{alphaMcmcPost-methods}
\alias{alphaMcmcPost}
\title{posterior estimates of the parameter alpha for each MCMC run}
\description{Posterior estimates of the parameter alpha for each MCMC run.}
\value{\code{alphaMcmcPost-methods}: \code{matrix}: Posterior estimates of the parameter alpha for each MCMC run.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\alias{alphaMcmcPost,MgsaMcmcResults-method}
\alias{alphaMcmcPost}
\usage{\S4method{alphaMcmcPost}{MgsaMcmcResults}(x)
}

