\name{betaPost-methods}
\alias{betaPost-methods}
\alias{betaPost}
\title{Posterior for beta}
\description{Realization values, posterior estimate and standard error for the parameter beta.}
\value{\code{betaPost-methods}: \code{data.frame}: realization values, posterior estimate and standard error for the parameter beta.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\alias{betaPost,MgsaResults-method}
\alias{betaPost}
\usage{\S4method{betaPost}{MgsaResults}(x)
}

