\name{CreateMgsaGoSetsFromGAF}
\alias{CreateMgsaGoSetsFromGAF}
\title{CreateMgsaGoSetsFromGAF}
\usage{CreateMgsaGoSetsFromGAF(filename, gene.id.col=3, go.id.col=5,
    evidence.col=7, name.col=10)}
\description{Creates a MgsaGoSets using annotation provided by a file in GAF 1.0
format. The file specified by the filename may be gzip-compressed.}
\details{TODO:  provide support for evidence codes, choose a better name}
