\name{itemIndices-methods}
\alias{itemIndices-methods}
\alias{itemIndices}
\title{Item indices of a MgsaSets}
\description{Returns the indices corresponding to the items}
\value{\code{itemIndices-methods}: a \code{integer}: the item indices.

}
\arguments{\item{sets}{an instance of class \code{\linkS4class{MgsaSets}}.}
\item{items}{\code{character} or \code{numeric} the items of interest.}
}
\alias{itemIndices,MgsaSets,character-method}
\alias{itemIndices}
\usage{\S4method{itemIndices}{MgsaSets,character}(sets, items)
\S4method{itemIndices}{MgsaSets,numeric}(sets, items)
}
\alias{itemIndices,MgsaSets,numeric-method}
\alias{itemIndices}

