\name{length-methods}
\alias{length,MgsaSets-method}
\alias{length}
\title{Length of a MgsaSets.}
\usage{\S4method{length}{MgsaSets}(x)
}
\description{Length (number of sets) of \code{\linkS4class{MgsaSets}}.}
\value{\code{integer} vector.}
\arguments{\item{x}{an instance of class \code{\linkS4class{MgsaSets}}.}
}

