% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgsa-package.R
\docType{package}
\name{mgsa-package}
\alias{mgsa-package}
\title{Model-based gene set analysis}
\description{
Model-based Gene Set Analysis (MGSA) is a Bayesian modeling
approach for gene set enrichment. The package mgsa implements MGSA and
tools to use MGSA together with the Gene Ontology.
}
\author{
Sebastian Bauer \email{Sebastian.Bauer@charite.de}, Julien Gagneur \email{julien.gagneur@embl.de}
}
\references{
S. Bauer, J. Gagneur and P. N. Robinson. GOing Bayesian: model-based gene set analysis of genome-scale data. Nucleic acids research, 2010.
}
\keyword{package}

