% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaGoSets-class.R
\docType{class}
\name{MgsaGoSets-class}
\alias{MgsaGoSets-class}
\title{Gene Ontology annotations}
\description{
This class represents gene ontology annotations.
}
\details{
For now, it is identical to the parental class \code{\linkS4class{MgsaSets}}.
}
\seealso{
\code{\link{readGAF}}
}

