\name{MgsaMcmcResults}
\alias{MgsaMcmcResults}
\title{MgsaMcmcResults}
\description{Instances of this class are used to hold the additional information 
that was provided by running an MCMC algorithm.}
\seealso{\code{\link{steps}}
\code{\link{restarts}}
\code{\link{alphaMcmcPost}}
\code{\link{betaMcmcPost}}
\code{\link{pMcmcPost}}}
