% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{nsamples}
\alias{nsamples}
\alias{nsamples,MgsaMcmcResults-method}
\title{How many samples per MCMC run collected}
\usage{
nsamples(x)

\S4method{nsamples}{MgsaMcmcResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
\code{integer}: how many samples per MCMC run collected.
}
\description{
how many samples collected per MCMC run.
}

