\name{plot,MgsaResults-method}
\alias{plot,MgsaResults-method}
\title{Plot method for MgsaResults objects}
\usage{
  \S4method{plot}{MgsaResults}(x, y, ...)
}
\arguments{
  \item{x}{a \code{\link{MgsaResults}}}

  \item{y}{unused}

  \item{...}{unused}
}
\description{
  Plot method for MgsaResults objects
}

