% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{plot,MgsaResults-method}
\alias{plot,MgsaResults-method}
\title{Plot method for MgsaResults objects}
\usage{
\S4method{plot}{MgsaResults}(x, y, ...)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaResults}}}

\item{y}{unused}

\item{...}{unused}
}
\description{
Plot method for MgsaResults objects
}

