% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{populationSize}
\alias{populationSize}
\alias{populationSize,MgsaResults-method}
\title{Size of the population of a MgsaResults}
\usage{
populationSize(x)

\S4method{populationSize}{MgsaResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\value{
\code{integer}: the size of the population.
}
\description{
The size of the population on which the analysis was run.
}

