\name{setAnnotations-methods}
\alias{setAnnotations-methods}
\alias{setAnnotations}
\title{Set annotations of a MgsaSets}
\description{Set annotations of a \code{\linkS4class{MgsaSets}}.}
\value{\code{setAnnotations-methods}: a \code{data.frame}: the set annotations.

}
\arguments{\item{sets}{an instance of class \code{\linkS4class{MgsaSets}}.}
\item{names}{\code{character} an optional vector specifying the names of interest.}
}
\alias{setAnnotations,MgsaSets,missing-method}
\alias{setAnnotations}
\usage{\S4method{setAnnotations}{MgsaSets,missing}(sets, names)
\S4method{setAnnotations}{MgsaSets,character}(sets, names)
}
\alias{setAnnotations,MgsaSets,character-method}
\alias{setAnnotations}

