\name{setsMcmcPost-methods}
\alias{setsMcmcPost-methods}
\alias{setsMcmcPost}
\title{posterior estimates of the the set marginal probabilities  for each MCMC run}
\description{Posterior estimates of the set marginal probabilities for each MCMC run.}
\value{\code{setsMcmcPost-methods}: \code{matrix}: Posterior estimates of the set marginal probabilities for each MCMC run.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\alias{setsMcmcPost,MgsaMcmcResults-method}
\alias{setsMcmcPost}
\usage{\S4method{setsMcmcPost}{MgsaMcmcResults}(x)
}

