\name{show,MgsaResults-method}
\alias{show,MgsaResults-method}
\title{Show an MgsaResults}
\usage{\S4method{show}{MgsaResults}(object)
}
\description{Show an \code{\linkS4class{MgsaResults}}.}
\value{an invisible \code{NULL}}
\arguments{\item{object}{an instance of class \code{\linkS4class{MgsaResults}}.}
}

