\name{show,MgsaSets-method}
\alias{show,MgsaSets-method}
\title{Show an MgsaSets}
\usage{
  \S4method{show}{MgsaSets}(object)
}
\arguments{
  \item{object}{an instance of class
  \code{\linkS4class{MgsaSets}}.}
}
\value{
  an invisible \code{NULL}
}
\description{
  Show an \code{\linkS4class{MgsaSets}}.
}

