% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{show,MgsaResults-method}
\alias{show,MgsaResults-method}
\title{Show an MgsaResults}
\usage{
\S4method{show}{MgsaResults}(object)
}
\arguments{
\item{object}{an instance of class \code{\linkS4class{MgsaResults}}.}
}
\value{
an invisible \code{NULL}
}
\description{
Show an \code{\linkS4class{MgsaResults}}.
}

