% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaSets-class.R
\docType{methods}
\name{show,MgsaSets-method}
\alias{show,MgsaSets-method}
\title{Show an MgsaSets}
\usage{
\S4method{show}{MgsaSets}(object)
}
\arguments{
\item{object}{an instance of class \code{\linkS4class{MgsaSets}}.}
}
\value{
an invisible \code{NULL}
}
\description{
Show an \code{\linkS4class{MgsaSets}}.
}

