\name{steps}
\alias{steps}
\alias{steps-methods}
\alias{steps,MgsaMcmcResults-method}
\title{How many steps per MCMC run}
\usage{
  steps(x)

  \S4method{steps}{MgsaMcmcResults}(x)
}
\arguments{
  \item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
  \code{integer}: how many steps per MCMC run.
}
\description{
  how many steps per MCMC run.
}

