\name{studySetSizeInPopulation-methods}
\alias{studySetSizeInPopulation-methods}
\alias{studySetSizeInPopulation}
\title{Size of the study set of a MgsaResults}
\description{The size of the study set on which the analysis was run.}
\value{\code{studySetSizeInPopulation-methods}: \code{integer}: the size of the study set.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\alias{studySetSizeInPopulation,MgsaResults-method}
\alias{studySetSizeInPopulation}
\usage{\S4method{studySetSizeInPopulation}{MgsaResults}(x)
}

