% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{alphaMcmcPost}
\alias{alphaMcmcPost}
\alias{alphaMcmcPost,MgsaMcmcResults-method}
\title{posterior estimates of the parameter alpha for each MCMC run}
\usage{
alphaMcmcPost(x)

\S4method{alphaMcmcPost}{MgsaMcmcResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
\code{matrix}: Posterior estimates of the parameter alpha for each MCMC run.
}
\description{
Posterior estimates of the parameter alpha for each MCMC run.
}

