\name{alphaPost-methods}
\alias{alphaPost-methods}
\alias{alphaPost}
\title{Posterior for alpha}
\description{Realization values, posterior estimate and standard error for the parameter alpha.}
\value{\code{alphaPost-methods}: \code{data.frame}: realization values, posterior estimate and standard error for the parameter alpha.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\alias{alphaPost,MgsaResults-method}
\alias{alphaPost}
\usage{\S4method{alphaPost}{MgsaResults}(x)
}

