\name{betaMcmcPost-methods}
\alias{betaMcmcPost-methods}
\alias{betaMcmcPost}
\title{posterior estimates of the parameter beta for each MCMC run}
\description{Posterior estimates of the parameter beta for each MCMC run.}
\value{\code{betaMcmcPost-methods}: \code{matrix}: Posterior estimates of the parameter beta for each MCMC run.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\alias{betaMcmcPost,MgsaMcmcResults-method}
\alias{betaMcmcPost}
\usage{\S4method{betaMcmcPost}{MgsaMcmcResults}(x)
}

