% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{betaMcmcPost}
\alias{betaMcmcPost}
\alias{betaMcmcPost,MgsaMcmcResults-method}
\title{posterior estimates of the parameter beta for each MCMC run}
\usage{
betaMcmcPost(x)

\S4method{betaMcmcPost}{MgsaMcmcResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
\code{matrix}: Posterior estimates of the parameter beta for each MCMC run.
}
\description{
Posterior estimates of the parameter beta for each MCMC run.
}

