\name{itemAnnotations-methods}
\alias{itemAnnotations-methods}
\alias{itemAnnotations}
\title{Item annotations of a MgsaSets}
\description{Item annotations of a \code{\linkS4class{MgsaSets}}.}
\value{\code{itemAnnotations-methods}: a \code{data.frame}: the item annotations.

}
\arguments{\item{sets}{an instance of class \code{\linkS4class{MgsaSets}}.}
\item{items}{\code{character} an optional vector specifying the items of interest.}
}
\alias{itemAnnotations,MgsaSets,missing-method}
\alias{itemAnnotations}
\usage{\S4method{itemAnnotations}{MgsaSets,missing}(sets, items)
\S4method{itemAnnotations}{MgsaSets,character}(sets, items)
}
\alias{itemAnnotations,MgsaSets,character-method}
\alias{itemAnnotations}

