\name{length,MgsaSets-method}
\alias{length,MgsaSets-method}
\title{Length of a MgsaSets.}
\usage{
  \S4method{length}{MgsaSets}(x)
}
\arguments{
  \item{x}{an instance of class
  \code{\linkS4class{MgsaSets}}.}
}
\value{
  \code{integer} vector.
}
\description{
  Length (number of sets) of \code{\linkS4class{MgsaSets}}.
}

