% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaSets-class.R
\docType{methods}
\name{length,MgsaSets-method}
\alias{length,MgsaSets-method}
\title{Length of a MgsaSets.}
\usage{
\S4method{length}{MgsaSets}(x)
}
\arguments{
\item{x}{an instance of class \code{\linkS4class{MgsaSets}}.}
}
\value{
\code{integer} vector.
}
\description{
Length (number of sets) of \code{\linkS4class{MgsaSets}}.
}

