\name{mgsa-package}
\alias{mgsa-package}
\alias{mgsa-package}
\title{Model-based gene set analysis}
\description{Model-based Gene Set Analysis (MGSA) is a Bayesian modeling
approach for gene set enrichment. The package mgsa implements MGSA and
tools to use MGSA together with the Gene Ontology.}
\docType{package}
\author{Sebastian Bauer \email{mail@sebastianbauer.info}, Julien Gagneur \email{julien.gagneur@embl.de}}
\keyword{package}
\references{S. Bauer, J. Gagneur and P. N. Robinson. GOing Bayesian: model-based gene set analysis of genome-scale data. Nucleic acids research, 2010.}
\keyword{package}

