\name{MgsaGoSets-class}
\alias{MgsaGoSets-class}
\alias{MgsaGoSets}
\title{Gene Ontology annotations}
\description{This class represents gene ontology annotations.}
\details{For now, it is identical to the parental class \code{\linkS4class{MgsaSets}}.}
\seealso{\code{\link{readGAF}}}
\section{Extends}{\code{\linkS4class{MgsaSets}}}

