\name{pMcmcPost-methods}
\alias{pMcmcPost-methods}
\alias{pMcmcPost}
\title{posterior estimates of the parameter p for each MCMC run}
\description{Posterior estimates of the parameter p for each MCMC run.}
\value{\code{pMcmcPost-methods}: \code{matrix}: Posterior estimates of the parameter p for each MCMC run.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\alias{pMcmcPost,MgsaMcmcResults-method}
\alias{pMcmcPost}
\usage{\S4method{pMcmcPost}{MgsaMcmcResults}(x)
}

