\name{populationSize-methods}
\alias{populationSize-methods}
\alias{populationSize}
\title{Size of the population of a MgsaResults}
\description{The size of the population on which the analysis was run.}
\value{\code{populationSize-methods}: \code{integer}: the size of the population.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\alias{populationSize,MgsaResults-method}
\alias{populationSize}
\usage{\S4method{populationSize}{MgsaResults}(x)
}

