\name{pPost-methods}
\alias{pPost-methods}
\alias{pPost}
\title{Posterior for beta}
\description{Realization values, posterior estimate and standard error for the parameter p.}
\value{\code{pPost-methods}: \code{data.frame}: realization values, posterior estimate and standard error for the parameter p.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\alias{pPost,MgsaResults-method}
\alias{pPost}
\usage{\S4method{pPost}{MgsaResults}(x)
}

