% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{pPost}
\alias{pPost}
\alias{pPost,MgsaResults-method}
\title{Posterior for beta}
\usage{
pPost(x)

\S4method{pPost}{MgsaResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\value{
\code{data.frame}: realization values, posterior estimate and standard error for the parameter p.
}
\description{
Realization values, posterior estimate and standard error for the parameter p.
}

