% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{restarts}
\alias{restarts}
\alias{restarts,MgsaMcmcResults-method}
\title{How many MCMC runs}
\usage{
restarts(x)

\S4method{restarts}{MgsaMcmcResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
\code{integer}: how many MCMC runs.
}
\description{
how many MCMC runs.
}

