\name{restarts-methods}
\alias{restarts-methods}
\alias{restarts}
\title{How many MCMC runs}
\description{how many MCMC runs.}
\value{\code{restarts-methods}: \code{integer}: how many MCMC runs.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\alias{restarts,MgsaMcmcResults-method}
\alias{restarts}
\usage{\S4method{restarts}{MgsaMcmcResults}(x)
}

