% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{setsMcmcPost}
\alias{setsMcmcPost}
\alias{setsMcmcPost,MgsaMcmcResults-method}
\title{posterior estimates of the the set marginal probabilities  for each MCMC run}
\usage{
setsMcmcPost(x)

\S4method{setsMcmcPost}{MgsaMcmcResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
\code{matrix}: Posterior estimates of the set marginal probabilities for each MCMC run.
}
\description{
Posterior estimates of the set marginal probabilities for each MCMC run.
}

