\name{setsResults-methods}
\alias{setsResults-methods}
\alias{setsResults}
\title{Posterior for each set}
\description{Number of items of the set in the population, the number of items of the set in the study set, the estimate of its marginal posterior probability and its standard error.}
\value{\code{setsResults-methods}: \code{data.frame}: For each set, number of items of the set in the population, number of items of the set in the study set, estimate of its marginal posterior probability and standard error.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\alias{setsResults,MgsaResults-method}
\alias{setsResults}
\usage{\S4method{setsResults}{MgsaResults}(x)
}

