% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{steps}
\alias{steps}
\alias{steps,MgsaMcmcResults-method}
\title{How many steps per MCMC run}
\usage{
steps(x)

\S4method{steps}{MgsaMcmcResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\value{
\code{integer}: how many steps per MCMC run.
}
\description{
how many steps per MCMC run.
}

