\name{steps-methods}
\alias{steps-methods}
\alias{steps}
\title{How many steps per MCMC run}
\description{how many steps per MCMC run.}
\value{\code{steps-methods}: \code{integer}: how many steps per MCMC run.

}
\arguments{\item{x}{a \code{\linkS4class{MgsaMcmcResults}}.}
}
\alias{steps,MgsaMcmcResults-method}
\alias{steps}
\usage{\S4method{steps}{MgsaMcmcResults}(x)
}

