% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MgsaResults-class.R
\docType{methods}
\name{studySetSizeInPopulation}
\alias{studySetSizeInPopulation}
\alias{studySetSizeInPopulation,MgsaResults-method}
\title{Size of the study set of a MgsaResults}
\usage{
studySetSizeInPopulation(x)

\S4method{studySetSizeInPopulation}{MgsaResults}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{MgsaResults}}.}
}
\value{
\code{integer}: the size of the study set.
}
\description{
The size of the study set on which the analysis was run.
}

